raddrinfo.c: suppress warnings
* ext/socket/raddrinfo.c (rsock_inspect_sockaddr): cast down
  explicitly to suppress warnings.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@40648 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/socket/raddrinfo.c
@@ -1189,7 +1189,7 @@ rsock_inspect_sockaddr(struct sockaddr *sockaddr_arg, socklen_t socklen, VALUE r
                 np = addr->sdl_data;
                 nlen = addr->sdl_nlen;
                 if (endp - np < nlen)
-                    nlen = endp - np;
+                    nlen = (int)(endp - np);
             }
             off = addr->sdl_nlen;
 
@@ -1197,7 +1197,7 @@ rsock_inspect_sockaddr(struct sockaddr *sockaddr_arg, socklen_t socklen, VALUE r
                 ap = addr->sdl_data + off;
                 alen = addr->sdl_alen;
                 if (endp - ap < alen)
-                    alen = endp - ap;
+                    alen = (int)(endp - ap);
             }
 
 	    CATSEP;
